red_thread/docker/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
matplotlib
tqdm
